requirements.txt
apache-airflow
requests
google-re2
msgspec==0.16.0